inference/requirements.txt
fastapi>=0.111.0
uvicorn>=0.29.0
python-multipart>=0.0.9
torch>=2.2.0
transformers>=4.56.0,<5.3.0,!=5.0.*,!=5.1.*
huggingface_hub>=0.23.0
soundfile>=0.12.1
librosa>=0.10.0
sentencepiece>=0.2.0
protobuf>=4.25.0
psutil>=5.9.0
faster-whisper>=1.0.0
sounddevice>=0.4.6
pynput>=1.7.0
pystray>=0.19.0
Pillow>=10.0.0
pyperclip>=1.8.0
customtkinter>=5.2.0
